Initialize the backing user and service objects.
Use the current values in the text input fields at page load time to initialize
the user and service objects that are used as inputs to the get_costs function.
This way, we avoid "NaN" and similar errors on a soft page reload, whewre the
browser retains the last-used values in the text input fields.