Embed the list of services into the index.html as compile-time.
Instead of loading the services.json array at page load time, use a Mustache template to write the service names and values into the index.html file at compile time. Also moved the source files into a src directory, and updated the docs to be copies of the files in the build directory.